requirements.txt
amqp==2.4.2
asn1crypto==0.24.0	
base58==1.0.3	
base58check==1.0.2	
billiard==3.5.0.5	
bitcoin==1.1.42	
blockchain==1.4.4	
celery==4.2.1	
cffi==1.12.2	
chardet==3.0.4	
coinbase==2.1.0	
confusable-homoglyphs==3.2.0	
cryptography==2.5	
django-celery-results==1.0.4	
django-cors-headers==2.5.2	
django-crispy-forms==1.7.2	
django-paypal==0.5.0	
django-registration==3.0	
ecdsa==0.13	
enum-compat==0.0.2	
exchangerates==0.2.0	
fft==0.1
future==0.17.1	
idna==2.8	
include==0.2.2	
json-rpc==1.12.1	
kombu==4.4.0	
lxml==3.7.3	
numpy==1.16.2	
paypalrestsdk==1.13.1	
pip==10.0.1	
pyOpenSSL==19.0.0	
pycryptodome==3.8.0	
python-bitcoinlib==0.10.1	
python-bitcoinrpc==1.0	
requests==2.13.0	
scipy==1.2.1	
setuptools==39.1.0	
simplejson==3.16.0	
six==1.11.0	
stripe==2.21.0	
urllib3==1.24.1	
utils==0.9.0	
vine==1.2.0
